requirements.txt
PyYAML>=6.0,<7.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
datasets>=3.0,<4.0
numpy>=1.24,<3.0
transformers>=4.44,<5.0
torch>=2.2,<3.0
accelerate>=0.34,<2.0
evaluate>=0.4,<1.0
rouge_score>=0.1.2,<1.0
sentencepiece>=0.1.99,<1.0
peft>=0.12,<1.0
sklearn-crfsuite>=0.5,<1.0
pulp>=2.8,<3.0
spacy>=3.7,<4.0
scispacy>=0.5.4,<1.0
bert-score>=0.3.13,<1.0
sacrebleu>=2.3,<3.0
tqdm>=4.65,<5.0
matplotlib>=3.7,<4.0
